Move the parameters package a level up
- This is done as it is semantically more correct to be in it's own package and not the connection package
